Simplified mDNS protocol.
Before we had few TXT records. These are had to get right at the IoT
side. So the solution there is to get rid of them and stick to the
single `zazanet` service name with 0 TXT records. All the rest via the
`GET /version` and endpoints like `GET /api/v1/services?type=logs`.